refactor: decouple Vue from core codegen, unify security types, remove dead abstractions
Architecture improvements identified through systematic review:

- Split client.ts (665→242 LOC): extract subscription-extractor.ts (pure GraphQL)
  and vue-subscription-builder.ts (Vue composables) from framework-agnostic core
- Unify SecurityConfig: delete duplicate in nitro/types, re-export from core
  (CoreSecurityConfig is now the single source of truth)
- Remove dead ScanAdapter interface and implementation: scanning goes through
  core functions directly, the adapter layer was designed but never wired up
- Type virtual module stubs: replace `any` with proper IResolvers, DebugInfoStub,
  and directive types for build-time type safety
- Remove `(i: any)` casts from debug route (types now flow from stubs)
