+++ src/core/types/adapter.ts
@@ -3,7 +3,7 @@
  * Defines the interface for framework-specific adapters
  */
 
-import type { CoreConfig, CoreContext, CoreLogger, ScanContext, ScannedResolver, ScanResult } from './index'
+import type { CoreConfig, CoreContext, CoreLogger, ScanContext } from './index'
 
 /**
  * Framework adapter interface
@@ -25,44 +25,3 @@ export interface FrameworkAdapter<TFramework = unknown> {
   /** Get framework logger adapted to CoreLogger interface */
   getLogger: (framework: TFramework) => CoreLogger
 }
-
-/**
- * Scan adapter interface
- * Provides high-level scanning operations using the framework adapter
- */
-export interface ScanAdapter<TFramework = unknown> {
-  /** Scan for GraphQL schema files */
-  scanSchemas: (framework: TFramework) => Promise<ScanResult<string>>
-
-  /** Scan for GraphQL files (schemas + operations) */
-  scanGraphql: (framework: TFramework) => Promise<ScanResult<string>>
-
-  /** Scan for resolver files */
-  scanResolvers: (framework: TFramework) => Promise<ScanResult<ScannedResolver>>
-
-  /** Scan for directive files */
-  scanDirectives: (framework: TFramework) => Promise<ScanResult<ScannedResolver>>
-
-  /** Scan for client documents */
-  scanDocuments: (framework: TFramework) => Promise<ScanResult<string>>
-}
-
-/**
- * Codegen adapter interface
- * Provides high-level codegen operations using the framework adapter
- */
-export interface CodegenAdapter<TFramework = unknown> {
-  /** Generate server types */
-  generateServerTypes: (framework: TFramework, options?: { silent?: boolean }) => Promise<void>
-
-  /** Generate client types */
-  generateClientTypes: (framework: TFramework, options?: { silent?: boolean }) => Promise<void>
-}
-
-/**
- * Combined full adapter interface
- */
-export interface FullFrameworkAdapter<TFramework = unknown>
-  extends FrameworkAdapter<TFramework>,
-  ScanAdapter<TFramework>,
-  CodegenAdapter<TFramework> {}
