+++ src/nitro/virtual/stubs.ts
@@ -4,14 +4,16 @@
  * At runtime, Nitro overrides these with actual virtual modules.
  */
 
+import type { IResolvers } from '@graphql-tools/utils'
+
 // #nitro-graphql/server-schemas
 export const schemas: Array<{ def: string }> = []
 
 // #nitro-graphql/server-resolvers
-export const resolvers: Array<{ resolver: any }> = []
+export const resolvers: Array<{ resolver: IResolvers }> = []
 
 // #nitro-graphql/server-directives
-export const directives: Array<{ directive: any }> = []
+export const directives: Array<{ directive: { name: string, locations: string[], transformer?: (schema: unknown) => unknown } }> = []
 
 // #nitro-graphql/module-config
 export const moduleConfig = {}
@@ -21,7 +23,25 @@ const importedConfig = {}
 export { importedConfig }
 
 // #nitro-graphql/debug-info
-export const debugInfo: Record<string, any> = {
+export interface DebugInfoStub {
+  isDev: boolean
+  framework: string
+  graphqlFramework?: string
+  federation?: Record<string, unknown>
+  scanned: {
+    schemas: number
+    schemaFiles: string[]
+    resolvers: number
+    resolverFiles: Array<{ specifier: string, imports: Array<{ name: string, type: string, as?: string }> }>
+    directives: number
+    directiveFiles: Array<{ specifier: string, imports: Array<{ name: string, type: string, as?: string }> }>
+    documents: number
+    documentFiles: string[]
+  }
+  virtualModules: Record<string, string>
+}
+
+export const debugInfo: DebugInfoStub = {
   isDev: false,
   framework: '',
   graphqlFramework: '',
@@ -40,4 +60,4 @@ export const debugInfo: Record<string, any> = {
 }
 
 // #nitro-graphql/pubsub
-export const pubsub: any = null
+export const pubsub: null = null
